refactor: Extract loadImage function to file
