+++ packages/insight-viewer/src/hooks/useImage/index.ts
@@ -1,53 +1,14 @@
 import { useEffect, useReducer } from 'react'
 import { LOADER_TYPE, LOADING_STATE, CONFIG } from '../../const'
 import { LoadingState, LoaderType } from '../../types'
-import {
-  loadImage as cornerstoneLoadImage,
-  CornerstoneImage,
-} from '../../utils/cornerstoneHelper'
-import { getHttpClient } from '../../utils/httpClient'
-import { formatError } from '../../utils/common'
+import { CornerstoneImage } from '../../utils/cornerstoneHelper'
 import { useImageLoader } from '../useImageLoader'
 import {
   imageLoadReducer,
   INITIAL_IMAGE_LOAD_STATE,
 } from '../../stores/imageLoadReducer'
-import { ImageLoad, DefaultGetImage, GetImage } from './types'
-
-const _getImage: DefaultGetImage = async ({ imageId, requestInterceptor }) => {
-  try {
-    return await cornerstoneLoadImage(imageId, {
-      loader: getHttpClient(requestInterceptor),
-    })
-  } catch (e) {
-    throw formatError(e)
-  }
-}
-
-/**
- * If successful, return cornerstone image.
- * If not successful, throw error.
- * getImage is pluggable for unit test.
- */
-export async function loadImage({
-  imageId,
-  requestInterceptor,
-  onError,
-  getImage = _getImage,
-}: Required<ImageLoad> & {
-  getImage?: GetImage
-  imageId: string
-}): Promise<CornerstoneImage> {
-  try {
-    return await getImage({
-      imageId,
-      requestInterceptor,
-    })
-  } catch (e) {
-    onError(e)
-    throw e
-  }
-}
+import { ImageLoad } from './types'
+import { loadImage } from './loadImage'
 
 export function useImage({
   requestInterceptor = CONFIG.requestInterceptor,
